+++ apps/backend/src/libs/enums/relation-name.enum.ts
@@ -12,6 +12,7 @@ const RelationName = {
 	TASKS: "tasks",
 	USER_DETAILS: "userDetails",
 	USER_TASK_DAYS: "userTaskDays",
+	USER_TASKS: "userTasks",
 	USERS: "users",
 } as const;