+++ packages/shared/src/modules/users/libs/types/user-update-password-form-dto.type.ts
@@ -1,0 +1,7 @@
+import { type UserUpdatePasswordRequestDto } from "./user-update-password-request-dto.type.js";
+
+type UserUpdatePasswordFormDto = {
+	confirmNewPassword: string;
+} & Omit<UserUpdatePasswordRequestDto, "email">;
+
+export { type UserUpdatePasswordFormDto };