+++ packages/shared/src/modules/users/libs/types/user-update-password-request-dto.type.ts
@@ -1,0 +1,6 @@
+type UserUpdatePasswordRequestDto = {
+	currentPassword: string;
+	newPassword: string;
+};
+
+export { type UserUpdatePasswordRequestDto };